+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/)
 - Added CrococosmoImporter to Analysis
 - Added type, dirs, name to CSVExporter
 - Invert height of building checkbox
+- Aggregate multiple maps in visualization #110
 - Auto Focus selected map part
 - Timmer added to applySettings in SettingsService